+++ Sourcery/Models/Method.swift
@@ -19,6 +19,9 @@ final class MethodParameter: NSObject, SourceryModel, Typed, Annotated {
 
     var typeAttributes: [String: Attribute] { return typeName.attributes }
 
+    /// Method parameter default value expression
+    var defaultValue: String?
+
     /// Annotations, that were created with // sourcery: annotation1, other = "annotation value", alterantive = 2
     var annotations: [String: NSObject] = [:]
 
@@ -26,19 +29,21 @@ final class MethodParameter: NSObject, SourceryModel, Typed, Annotated {
     // sourcery: skipEquality, skipDescription, skipCoding, skipJSExport
     internal var __parserData: Any?
 
-    init(argumentLabel: String?, name: String = "", typeName: TypeName, type: Type? = nil, annotations: [String: NSObject] = [:]) {
+    init(argumentLabel: String?, name: String = "", typeName: TypeName, type: Type? = nil, defaultValue: String? = nil, annotations: [String: NSObject] = [:]) {
         self.typeName = typeName
         self.argumentLabel = argumentLabel
         self.name = name
         self.type = type
+        self.defaultValue = defaultValue
         self.annotations = annotations
     }
 
-    init(name: String = "", typeName: TypeName, type: Type? = nil, annotations: [String: NSObject] = [:]) {
+    init(name: String = "", typeName: TypeName, type: Type? = nil, defaultValue: String? = nil, annotations: [String: NSObject] = [:]) {
         self.typeName = typeName
         self.argumentLabel = name
         self.name = name
         self.type = type
+        self.defaultValue = defaultValue
         self.annotations = annotations
     }
 
@@ -48,6 +53,7 @@ final class MethodParameter: NSObject, SourceryModel, Typed, Annotated {
             guard let name: String = aDecoder.decode(forKey: "name") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["name"])); fatalError() }; self.name = name
             guard let typeName: TypeName = aDecoder.decode(forKey: "typeName") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["typeName"])); fatalError() }; self.typeName = typeName
             self.type = aDecoder.decode(forKey: "type")
+            self.defaultValue = aDecoder.decode(forKey: "defaultValue")
             guard let annotations: [String: NSObject] = aDecoder.decode(forKey: "annotations") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["annotations"])); fatalError() }; self.annotations = annotations
         }
 
@@ -56,6 +62,7 @@ final class MethodParameter: NSObject, SourceryModel, Typed, Annotated {
             aCoder.encode(self.name, forKey: "name")
             aCoder.encode(self.typeName, forKey: "typeName")
             aCoder.encode(self.type, forKey: "type")
+            aCoder.encode(self.defaultValue, forKey: "defaultValue")
             aCoder.encode(self.annotations, forKey: "annotations")
         }
         // sourcery:end